styling refactor
