+++ examples/examples.py
@@ -13,7 +13,7 @@ def example_1_create_star_chart():
         lon=-116.836394,
         dt=tz.localize(datetime.now().replace(hour=22)),
         limiting_magnitude=4.6,
-        style=sp.styles.BLUE,
+        style=sp.styles.ZENITH_BLUE_MEDIUM,
         resolution=2000,
     )
     p.export("01_star_chart.png")
@@ -31,7 +31,7 @@ def example_2_create_star_chart_extra():
         lon=-117.038611,
         dt=tz.localize(datetime.now().replace(hour=22)),
         limiting_magnitude=4.6,
-        style=sp.styles.GRAYSCALE,
+        style=sp.styles.ZENITH_GRAYSCALE,
         resolution=2000,
     )
     p.plot_object(
@@ -50,7 +50,7 @@ def example_2_create_star_chart_extra():
 def example_3_create_map_orion():
     import starplot as sp
 
-    style = sp.styles.MAP_BLUE.extend(
+    style = sp.styles.MAP_BLUE_LIGHT.extend(
         {
             "bayer_labels": {
                 "font_name": "GFS Didot",  # use a better font for Greek letters
@@ -100,7 +100,7 @@ def example_3_create_map_orion():
 def example_style():
     import starplot as sp
 
-    sp.styles.MAP_BLUE.dump_to_file("blue.yml")
+    sp.styles.MAP_BLUE_LIGHT.dump_to_file("map_blue_light.yml")
 
 
 # Documented Examples
