+++ src/starplot/styles/__init__.py
@@ -1,0 +1,4 @@
+from .base import *  # noqa: F401
+from .helpers import *  # noqa: F401
+
+# from .extensions import *  # noqa: F401
